requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
SQLAlchemy==2.0.30
sqlmodel==0.0.22
pydantic[email]==2.8.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pyjwt==2.9.0
python-dotenv==1.0.1
pypdf==5.0.1
transformers==4.42.4
torch==2.3.1
sentence-transformers==3.0.1
sentencepiece==0.2.0

